fix(files): handle indestructible in api/files#destroy

+++ app/controllers/concerns/folio/console/api/file_controller_base.rb
@@ -55,8 +55,24 @@ module Folio::Console::Api::FileControllerBase
   end
 
   def destroy
+    indestructible_reason = folio_console_record.indestructible_reason
+    indestructible_reason ||= if folio_console_record.file_placements.exists?
+      I18n.t("folio.file.cannot_destroy_file_with_placements")
+    end
+
+    if indestructible_reason
+      render json: { errors: [
+        status: 422,
+        title: "ActiveRecord::RecordNotDestroyed",
+        detail: indestructible_reason,
+      ] }, status: 422
+      return
+    end
+
     folio_console_record.destroy!
     render json: { status: 200 }
+  rescue ActiveRecord::RecordNotDestroyed => e
+    render_error(e, status: 422)
   end
 
   def pagination
